Fixed Logout

+++ TCSlackbot.Logic/Utils/SecretManager.cs
@@ -1,6 +1,10 @@
 ﻿using Microsoft.Azure.KeyVault;
+using Microsoft.Azure.KeyVault.Models;
+
 using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.Extensions.Configuration;
+using System;
+using System.Threading.Tasks;
 
 namespace TCSlackbot.Logic
 {
@@ -24,13 +28,23 @@ namespace TCSlackbot.Logic
             _configuration[key] = value;
         }
 
-        public async System.Threading.Tasks.Task DeleteSecretAsync(string key)
+        public async Task DeleteSecretAsync(string key)
         {
             var azureServiceTokenProvider = new AzureServiceTokenProvider();
             var keyVaultClient = new KeyVaultClient(new KeyVaultClient.AuthenticationCallback(azureServiceTokenProvider.KeyVaultTokenCallback));
+            if( _configuration[key] == null )
+            {
+                return;
+            }
+            try
+            {
+                await keyVaultClient.DeleteSecretAsync(KeyVaultEndpoint, key);
 
-            await keyVaultClient.DeleteSecretAsync(KeyVaultEndpoint, key);
-
+            }
+            catch(KeyVaultErrorException ex)
+            {
+                Console.WriteLine("\n\nError:" + ex.Message);
+            }
             // Reload the configuration because we added a new secret
             ((IConfigurationRoot)_configuration).Reload();
         }
